backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.1
Django==5.1.5
greenlet==3.1.1
jwt==1.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
typing_extensions==4.12.2

tabulate~=0.9.0